+++ x/incentivization/types/keys.go
@@ -1,6 +1,0 @@
-package types
-
-const (
-	ModuleName = "incentivization"
-	StoreKey   = ModuleName
-)